Update winit dependency from 0.29 to 0.30
Mostly done via Claude Sonnet 4.5.

Prompt:
```
Convert this @main.rs to use the winit 0.30 version. This should now use
the new ApplicationHandler trait.
```

A few tweaks needed afterwards:
* Frontend object is now relocated, but we pass in the C midend pointers
  to our FFI interfaces, which can't move.  So, wrap the Frontend in a
  `Box` to force it onto the heap and prevent things from relocating.

* The frontend.tick() was incorrectly being called from the window_event
  handler, not the about_to_wait.

It also seems to have dropped the usage of the winit_input_helper, which
looking at the code doesn't seem to be a huge loss, and is maybe better.
Will need to reconsider whether adding it back is worthwhile.